+++ NEWS.md
@@ -1,5 +1,11 @@
 **Please check the [latest news (change log)](https://psychbruce.github.io/PsychWordVec/news/index.html) and keep this package updated.**
 
+# PsychWordVec 0.0.7 (Aug 2022)
+
+## New Features
+
+-   New `plot_wordvecs_tSNE()` function: Visualize 2-D or 3-D word vectors with dimensionality reduced using the t-Distributed Stochastic Neighbor Embedding (t-SNE) method.
+
 # PsychWordVec 0.0.6 (Jul 2022)
 
 ## New Features